fix: not converting array of strings values
Fix not converting array of strings values. This was likely caused by naming the variable for attribute `value` and the one for the attribute value `key`. Change the variable names to be more accurate.

Add a better JSDoc type annotation for `this.args` to make more obvious when errors in normalizing attribute values occur.